+++ gui/admin/src/lib/_typemap.js
@@ -6,4 +6,5 @@ export default {
   float: 'value_float',
   integer: 'value_int',
   string: 'value',
+  text: 'value'
 };