+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 4.4.20 - 15 July 2020 - Admin improvements
+* Correctly display arrays in fields
+* Show `text` fields as textarea
+* Fix updating `text` fields
+* Add string filters to text fields
+* Correctly display values in quotes, square brackets, etc. in edit view
+* Improve example hints
+
 ## 4.4.18 - 14 July 2020
 * Update GraphiQL to 1.x